requirements.txt
streamlit
transformers
huggingface_hub
requests
sseclient-py
